worked on com player, overall datapath
in which jules loses her mind screaming at wires connecting when they
shouldn't. dear god this is hell and there's no automatic way to
organize a schematic altera confirmed to be satan.

testing needed for the computer and dealer circuits (along with the
datapath ofc), but i'm going to do most of that on the FPGA because i
can't tell what's modelsim being weird and what's an actual issue.